refactor(service): split Crypto service into files
