+++ src/extension/background-script/CryptoServices/getMyProveBio.ts
@@ -1,0 +1,13 @@
+import { encodeArrayBuffer } from '../../../utils/type-transform/String-ArrayBuffer'
+import { toCompressSecp256k1Point } from '../../../utils/type-transform/SECP256k1-Compression'
+import { queryMyIdentityAtDB } from '../../../database/people'
+import { PersonIdentifier } from '../../../database/type'
+//#endregion
+//#region ProvePost, create & verify
+export async function getMyProveBio(whoAmI: PersonIdentifier): Promise<string | null> {
+    const myIdentity = await queryMyIdentityAtDB(whoAmI)
+    if (!myIdentity) return null
+    const pub = await crypto.subtle.exportKey('jwk', myIdentity.publicKey)
+    const compressed = toCompressSecp256k1Point(pub.x!, pub.y!)
+    return `🔒${encodeArrayBuffer(compressed)}🔒`
+}
